# 03_openapi_original.yaml
openapi: 3.0.3
info:
  title: Kaveri Stays API
  version: 1.0.0
  description: API for Kaveri Stays
servers:
  - url: http://127.0.0.1:8000
...
paths:
...
  /bookings:
    get:
      tags: [Bookings]
      summary: List bookings
      parameters:
        - name: property_id
          in: query
          schema: {type: integer}
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/BookingStatus'
        - name: date_from
          in: query
          schema: {type: string, format: date}
        - name: date_to
          in: query
          schema: {type: string, format: date}
        - name: guest_id
          in: query
          schema: {type: integer}
        - name: sort
          in: query
          schema:
            type: string
            enum: [booking_id, check_in, check_out, created_at, status]
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200': {description: Paginated bookings}
        '400': {description: Invalid filter or sort}
        '401': {description: Missing or invalid token}
        '403': {description: Access denied}

    post:
      tags: [Bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201': {description: Booking created}
        '400': {description: Invalid request}
        '401': {description: Missing or invalid token}
        '403': {description: Not allowed}
        '409': {description: Room unavailable or constraint conflict}
  /bookings/{booking_id}:
    get:
      tags: [Bookings]
      summary: Get booking
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200': {description: Booking details}
        '401': {description: Missing or invalid token}
        '403': {description: Access denied}
        '404': {description: Booking not found}

    patch:
      tags: [Bookings]
      summary: Change booking status
      parameters:
        - $ref: '#/components/parameters/BookingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeStatusRequest'
      responses:
        '200': {description: Status changed}
        '400': {description: Illegal transition}
        '401': {description: Missing or invalid token}
        '403': {description: Transition not allowed}
        '404': {description: Booking not found}
        '409': {description: State conflict}
  /bookings/{booking_id}/payments:
    post:
      tags: [Payments]
      summary: Add a payment
      parameters:
        - $ref: '#/components/parameters/BookingId'
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201': {description: Payment recorded}
        '400': {description: Invalid payment}
        '401': {description: Missing or invalid token}
        '403': {description: Not allowed}
        '404': {description: Booking not found}
        '409': {description: Payment conflict}
...
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    BookingId:
      name: booking_id
      in: path
      required: true
      schema: {type: integer, minimum: 1}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    Offset:
      name: offset
      in: query
      schema: {type: integer, minimum: 0, default: 0}
...
  schemas:
...
    BookingStatus:
      type: string
      enum: [pending, confirmed, checked_in, checked_out, cancelled]
    CreateBookingRequest:
      type: object
      required: [property_id, room_id, check_in, check_out, guest_count, deposit_amount]
      properties:
        property_id: {type: integer}
        room_id: {type: integer}
        check_in: {type: string, format: date}
        check_out: {type: string, format: date}
        guest_count: {type: integer, minimum: 1}
        deposit_amount: {type: number, minimum: 0}
    ChangeStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/BookingStatus'
    PaymentRequest:
      type: object
      required: [amount]
      properties:
        amount: {type: number, minimum: 0}
...